+++ changelog.md
@@ -1,8 +1,10 @@
 # Change Log
-
 ## 4.5.1
+**Features**
+ * Issue #900. Add `@ApiModelProperty` support to `elide-swagger` that makes it possible to customize `description`, `example`, `readOnly` and `required` attributes of object definitions in resulting generates Swagger document.
+
 **Fixes**
-[Security] Bump jackson databind from 2.9.9 to 2.9.9.3
+ * [Security] Bump jackson databind from 2.9.9 to 2.9.9.3
 
 ## 4.5.0
 **Features**